Encoder + SimpleEncoder
We have two interfaces:
* a simplified one that deals in inputstreams. Can either turn one into
a fountain_encoder_stream, and then use that to encode each image frame,
or skip that step and encode each image frame from the initial
inputstream.
* a file-based one that reads files and does the right thing for you.

To accomodate emscripten+wasm, these file functions are being split out
into a different file/class. I *think* the split is fairly logically
sound. (in any case, the alternative seems to be ifdefs :grimacing: )
