+++ src/lib/encoder/test/EncoderTest.cpp
@@ -1,7 +1,7 @@
 /* This code is subject to the terms of the Mozilla Public License, v.2.0. http://mozilla.org/MPL/2.0/. */
 #include "unittest.h"
 
-#include "encoder/FileEncoder.h"
+#include "encoder/Encoder.h"
 #include "fountain/FountainInit.h"
 #include "image_hash/average_hash.h"
 #include "serialize/format.h"
@@ -22,7 +22,7 @@ TEST_CASE( "EncoderTest/testVanilla", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.vanilla";
 
-	FileEncoder enc(40, 4, 2);
+	Encoder enc(40, 4, 2);
 	assertEquals( 3, enc.encode(inputFile, outPrefix) );
 
 	std::vector<uint64_t> hashes = {0xc47ced006f253562, 0x2d8cc44256ccb9eb, 0x6ed0efb800000000};
@@ -45,7 +45,7 @@ TEST_CASE( "EncoderTest/testFountain", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	FileEncoder enc(40, 4, 2);
+	Encoder enc(40, 4, 2);
 	assertEquals( 5, enc.encode_fountain(inputFile, outPrefix, 0) );
 
 	std::vector<uint64_t> hashes = {0xbb1cc62b662abfe5, 0xf586f6466a5b194, 0x8c2f0f40e6ecb08b, 0x93a3830d042966e1, 0x21f4ef48d3ed685e};
@@ -68,7 +68,7 @@ TEST_CASE( "EncoderTest/testFountain.Compress", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	FileEncoder enc(40, 4, 2);
+	Encoder enc(40, 4, 2);
 	assertEquals( 1, enc.encode_fountain(inputFile, outPrefix) );
 
 	uint64_t hash = 0xf8cde200e90582e4;
@@ -82,7 +82,7 @@ TEST_CASE( "EncoderTest/testPiecemealFountainEncoder", "[unit]" )
 	// use the fountain_encoder_stream directly on a char*,int pair
 	FountainInit::init();
 	MakeTempDirectory tempdir;
-	FileEncoder enc(40, 4, 2);
+	Encoder enc(40, 4, 2);
 
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string contents = File(inputFile).read_all();
